Adding common runtime 2 (#297)
